pyproject.toml
[tool.poetry]
name = "demo-feature-flag-service"
version = "0.1.0"
description = ""
authors = ["Pete Hodgson <git@thepete.net>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.2"
uvicorn = "^0.32.0"
opentelemetry-instrumentation-fastapi = "^0.48b0"
opentelemetry-api = "^1.27.0"
opentelemetry-sdk = "^1.27.0"
opentelemetry-exporter-otlp-proto-grpc = "^1.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
